TensorNetworkAD2

Documentation for TensorNetworkAD2.

TensorNetworkAD2.StopFunction

TensorNetworkAD2.fixedpoint

TensorNetworkAD2.trg_i

TensorNetworkAD2.StopFunction — Method

(st::StopFunction)(state)

stopfunction for ctmrg, returning true if singular values are converged or the maximum number of iterations is reached. Implemented as a closure since it needs to remember the last singular values it saw for comparison.

TensorNetworkAD2.fixedpoint — Method

fixedpoint(f, guess, stopfun)

return the result of applying guess = f(guess) until convergence. Convergence is decided by applying stopfun(guess) which returns a Boolean.

TensorNetworkAD2.trg_i — Method

trg(T::ITensor; χmax::Int, nsteps::Int) -> κ, T

Perform the TRG algorithm on the partition function composed of the ITensor T.

The indices of T must obey come in pairs (sₕ => sₕ') and `(sᵥ => sᵥ').

χmax is the maximum renormalized bond dimension.

nsteps are the number of renormalization steps performed.

The outputs are κ, the partition function per site, and the final renormalized ITensor T.